backend/requirements.txt
Flask
openai
psycopg2-binary
SQLAlchemy
pytest
python-dotenv
flask_sqlalchemy
flask_cors
google-cloud-speech
google-cloud
google-cloud-storage
google-cloud-videointelligence
assemblyai
google
textract==1.6.1
google-cloud-vision
flask_migrate
elasticsearch